Simplify adding new enrichments on run-time
